When using interestingness tests in autoBisect, sometimes cached shells do not have temp dirs. Thus remove the temp dir only if it exists.

+++ autobisect-js/autoBisect.py
@@ -304,7 +304,8 @@ def externalTestAndLabel(options, interestingness):
             innerResult = ('bad', 'interesting')
         else:
             innerResult = ('good', 'not interesting')
-        shutil.rmtree(tempPrefix + shell.getHgHash())
+        if os.path.isdir(tempPrefix + shell.getHgHash()):
+            shutil.rmtree(tempPrefix + shell.getHgHash())
         return innerResult
     return inner
